+++ games/highrise/src/maps/index.ts
@@ -2,11 +2,12 @@ import type { MapTheme } from './MapTheme'
 import { defaultMap } from './default'
 import { favelaMap } from './favela'
 import { nyMap } from './ny'
+import { arboreaMap } from './arborea'
 
 export type { MapTheme } from './MapTheme'
 
 /** Ordered list of all selectable maps, in the order they appear in the menu. */
-export const ALL_MAPS: MapTheme[] = [defaultMap, favelaMap, nyMap]
+export const ALL_MAPS: MapTheme[] = [defaultMap, favelaMap, nyMap, arboreaMap]
 
 export const DEFAULT_MAP_ID = defaultMap.id